rename simulated tests
