+++ src/software/ai/hl/stp/stp_test.cpp
@@ -9,14 +9,15 @@
 #include "software/test_util/test_util.h"
 #include "software/util/generic_factory/generic_factory.h"
 
-//class STPTest : public ::testing::Test
+// class STPTest : public ::testing::Test
 //{
 //   public:
 //    STPTest()
 //        : world(::TestUtil::createBlankTestingWorld()),
 //          mutable_ai_control_config(std::make_shared<AiControlConfig>()),
 //          ai_control_config(
-//              std::const_pointer_cast<const AiControlConfig>(mutable_ai_control_config)),
+//              std::const_pointer_cast<const
+//              AiControlConfig>(mutable_ai_control_config)),
 //          play_config(std::make_shared<const ThunderbotsConfig>()->getPlayConfig()),
 //          default_play_constructor([this]() -> std::unique_ptr<Play> {
 //              return std::make_unique<HaltTestPlay>(play_config);
@@ -41,17 +42,17 @@
 //    STP stp;
 //};
 //
-//TEST_F(STPTest, test_only_test_plays_are_registered_in_play_factory)
+// TEST_F(STPTest, test_only_test_plays_are_registered_in_play_factory)
 //{
-//    auto play_names = GenericFactory<std::string, Play, PlayConfig>::getRegisteredNames();
-//    EXPECT_EQ(2, play_names.size());
+//    auto play_names = GenericFactory<std::string, Play,
+//    PlayConfig>::getRegisteredNames(); EXPECT_EQ(2, play_names.size());
 //    EXPECT_EQ(std::count(play_names.begin(), play_names.end(), TYPENAME(MoveTestPlay)),
 //              1);
 //    EXPECT_EQ(std::count(play_names.begin(), play_names.end(), TYPENAME(HaltTestPlay)),
 //              1);
 //}
 //
-//TEST_F(STPTest, test_exception_thrown_when_no_play_applicable)
+// TEST_F(STPTest, test_exception_thrown_when_no_play_applicable)
 //{
 //    // Put the ball where both its x and y coordinates are negative. Neither test Play
 //    // is applicable in this case
@@ -60,7 +61,7 @@
 //    EXPECT_THROW(stp.calculateNewPlay(world), std::runtime_error);
 //}
 //
-//TEST_F(STPTest, test_calculate_new_play_when_one_play_valid)
+// TEST_F(STPTest, test_calculate_new_play_when_one_play_valid)
 //{
 //    // Only the HaltTestPlay should be applicable
 //    world = ::TestUtil::setBallPosition(world, Point(-1, 1), Timestamp::fromSeconds(0));
@@ -69,13 +70,15 @@
 //    EXPECT_EQ(objectTypeName(*play), TYPENAME(HaltTestPlay));
 //}
 //
-//TEST_F(STPTest, test_calculate_new_play_when_multiple_plays_valid)
+// TEST_F(STPTest, test_calculate_new_play_when_multiple_plays_valid)
 //{
 //    // Both HaltTestPlay and MoveTestPlay should be applicable
 //    world = ::TestUtil::setBallPosition(world, Point(1, 1), Timestamp::fromSeconds(0));
 //
-//    // We expect a random selection of the plays that are applicable. This test should be
-//    // deterministic because we have provided a seed for this test (in the setUp function)
+//    // We expect a random selection of the plays that are applicable. This test should
+//    be
+//    // deterministic because we have provided a seed for this test (in the setUp
+//    function)
 //
 //    std::unique_ptr<Play> play;
 //    std::vector<std::string> actual_play_names;
@@ -96,13 +99,13 @@
 //    EXPECT_EQ(expected_play_names, actual_play_names);
 //}
 //
-//TEST_F(STPTest, test_current_play_initially_unassigned)
+// TEST_F(STPTest, test_current_play_initially_unassigned)
 //{
 //    EXPECT_EQ(stp.getCurrentPlayName(), std::nullopt);
 //}
 //
 //// Test that we can successfully assign Plays when there is currently no Play assigned
-//TEST_F(STPTest, test_play_assignment_transition_from_unassigned_to_assigned)
+// TEST_F(STPTest, test_play_assignment_transition_from_unassigned_to_assigned)
 //{
 //    // Only the HaltTestPlay should be applicable
 //    world = ::TestUtil::setBallPosition(world, Point(-1, 1), Timestamp::fromSeconds(0));
@@ -110,7 +113,7 @@
 //    EXPECT_EQ(*(stp.getCurrentPlayName()), TYPENAME(HaltTestPlay));
 //}
 //
-//TEST_F(
+// TEST_F(
 //    STPTest,
 //    test_play_assignment_from_one_play_to_another_when_current_play_invariant_no_longer_holds)
 //{
@@ -119,15 +122,18 @@
 //    stp.getIntents(world);
 //    EXPECT_EQ(*(stp.getCurrentPlayName()), TYPENAME(HaltTestPlay));
 //
-//    // The HaltTestPlays invariant should no longer hold, and the MoveTestPlay should now
+//    // The HaltTestPlays invariant should no longer hold, and the MoveTestPlay should
+//    now
 //    // be applicable
 //    world = ::TestUtil::setBallPosition(
-//        world, world.field().enemyCornerNeg() + Vector(1, 0), Timestamp::fromSeconds(0));
+//        world, world.field().enemyCornerNeg() + Vector(1, 0),
+//        Timestamp::fromSeconds(0));
 //    stp.getIntents(world);
 //    EXPECT_EQ(*(stp.getCurrentPlayName()), TYPENAME(MoveTestPlay));
 //}
 //
-//TEST_F(STPTest, test_play_assignment_from_one_play_to_another_when_current_play_is_done)
+// TEST_F(STPTest,
+// test_play_assignment_from_one_play_to_another_when_current_play_is_done)
 //{
 //    // Only the MoveTestPlay should be applicable
 //    world = ::TestUtil::setBallPosition(world, Point(1, -1), Timestamp::fromSeconds(0));
@@ -141,9 +147,10 @@
 //    EXPECT_EQ(*(stp.getCurrentPlayName()), TYPENAME(HaltTestPlay));
 //}
 //
-//TEST_F(STPTest, test_fallback_play_assigned_when_no_new_plays_are_applicable)
+// TEST_F(STPTest, test_fallback_play_assigned_when_no_new_plays_are_applicable)
 //{
-//    // TODO: Update this test when https://github.com/UBC-Thunderbots/Software/issues/396
+//    // TODO: Update this test when
+//    https://github.com/UBC-Thunderbots/Software/issues/396
 //    // is completed. For now, we just check that the previous play remains assigned
 //
 //    // Only the HaltTestPlay should be applicable
@@ -153,12 +160,12 @@
 //
 //    // Put the ball where both its x and y coordinates are negative. Neither test Play
 //    // is applicable in this case
-//    world = ::TestUtil::setBallPosition(world, Point(-1, -1), Timestamp::fromSeconds(0));
-//    stp.getIntents(world);
+//    world = ::TestUtil::setBallPosition(world, Point(-1, -1),
+//    Timestamp::fromSeconds(0)); stp.getIntents(world);
 //    EXPECT_EQ(*(stp.getCurrentPlayName()), TYPENAME(HaltTestPlay));
 //}
 //
-//TEST_F(STPTest, test_get_play_info)
+// TEST_F(STPTest, test_get_play_info)
 //{
 //    // Only the HaltTestPlay should be applicable
 //    world = ::TestUtil::setBallPosition(world, Point(-1, 1), Timestamp::fromSeconds(0));
