+++ src/software/ai/hl/stp/tactic/attacker/attacker_tactic_shoot_goal_test.cpp
