+++ src/software/ai/hl/stp/tactic/dribble/dribble_tactic_test.cpp
@@ -1,8 +1,9 @@
+#include "software/ai/hl/stp/tactic/dribble/dribble_tactic.h"
+
 #include <gtest/gtest.h>
 
 #include <utility>
 
-#include "software/ai/hl/stp/tactic/dribble/dribble_tactic.h"
 #include "software/geom/algorithms/contains.h"
 #include "software/simulated_tests/non_terminating_validation_functions/robot_not_excessively_dribbling_validation.h"
 #include "software/simulated_tests/simulated_tactic_test_fixture.h"
