+++ src/software/ai/hl/stp/tactic/move/BUILD
@@ -34,16 +34,6 @@ cc_test(
     deps = [
         ":move_tactic",
         "//shared/test_util:tbots_gtest_main",
-        "//software/test_util",
-    ],
-)
-
-cc_test(
-    name = "simulated_move_tactic_test",
-    srcs = ["simulated_move_tactic_test.cpp"],
-    deps = [
-        ":move_tactic",
-        "//shared/test_util:tbots_gtest_main",
         "//software/simulated_tests:simulated_tactic_test_fixture",
         "//software/simulated_tests/terminating_validation_functions",
         "//software/simulated_tests/validation:validation_function",
