+++ src/software/ai/hl/stp/tactic/move_goalie_to_goal_line/move_goalie_to_goal_line_tactic_test.cpp
@@ -1,6 +1,7 @@
+#include "software/ai/hl/stp/tactic/move_goalie_to_goal_line/move_goalie_to_goal_line_tactic.h"
+
 #include <gtest/gtest.h>
 
-#include "software/ai/hl/stp/tactic/move_goalie_to_goal_line/move_goalie_to_goal_line_tactic.h"
 #include "software/geom/algorithms/contains.h"
 #include "software/simulated_tests/simulated_tactic_test_fixture.h"
 #include "software/simulated_tests/terminating_validation_functions/robot_state_validation.h"
